Create Controllers & Refactor Dashboard
