+++ server/src/controllers/seasonController.js
@@ -1,0 +1,19 @@
+const axios = require("axios");
+const apiKey = process.env.FB_API_KEY;
+const apiHost = process.env.FB_API_HOST;
+const apiBaseUrl = process.env.FB_API_URL;
+
+async function updateSeasons() {
+  try {
+    const options = {
+      method: 'GET',
+      url: apiBaseUrl + '/seasons',
+    }
+  } catch (error) {
+    console.log('Erreur lors de la mise à jour des données:', error);
+  }
+}
+
+module.exports = {
+  updateSeasons,
+};
