+++ server/src/services/betService.js
@@ -1,0 +1,7 @@
+function calculatePoints(wins, draws, loses) {
+  return (wins * 3) + draws;
+}
+
+module.exports = {
+  calculatePoints,
+};
